requirements.txt
python-dotenv
selenium
openai
sqlalchemy
psycopg2